Added some query optimisation strategies (filter and join order optimisations). At the moment these two strategies will always be applied on the generated query.
For future versions it is planed to create the queries using sesame functionality. That means we will use Objects to create queries instead of simply writing the queries as strings. This will allow us to switch to sesame and get rid of the jena query optimizer. Besides that it is planed to let the user configure what strategies should be applied and which not. Furthermore there will be the possibility to register own (sesame) optimisation strategies.