Altered the ObjectParser to have a clear parameter at the parse() method.
Setting the clear parameter to true invokes standard behaviour of clearing the parsed Annotations out of the local Anno4j instance for performance issues.
This however causes the direct Annotations to not be printable. This can be counteracted by setting clear to false. In this case, the performance issues (see issue #131) might occur.